PRO-440 FIX Fetch plan in apply step
Was in the preparation of a workflow run, however not all credentials
may be created by then, so push the fetching of the plan as far back
as possible to the apply step.
